ci: add `node-launchpad` binary to release process
The new TUI for the node manager is a new, separate binary that is being added in to the release
process. It will have a Github release.

For some reason it seems to have became necessary to explicitly call `rustup target` during the
artifacts build process. Without doing it, I was getting build errors on macOS.